ViewPager with PagerTitleStrip
This is not an implementation of tabs, but a slight addition to the
"swipe views only" variant in branch 01_swipe_views_and_actionbar.

PagerTitleStrip is a title bar for the ViewPager which displays the title
of the currently selected page, as well as the previous and next page.

Unlike ActionBar tabs, the PagerTitleStrip is not a part of the ActionBar,
but rather a part of the ViewPager. Furthermore, unlike tabs, the
PagerTitleStrip is non-interactive, i.e. it cannot be clicked upon.

To implement a PagerTitleStrip in addition to pure swipe views, the
following is necessary:

- Add a PagerTitleStrip element as a child of the ViewPager element in
  the layout XML
- Make sure to override the getPageTitle method of the PagerAdapter

That's all.

References:
https://developer.android.com/training/implementing-navigation/lateral.html#PagerTitleStrip